Use nativeDebugCallback to print C++ debug in java. Helps on Windows CE
git-svn-id: http://bluecove.googlecode.com/svn/trunk@193 3cac39a0-5d1b-0410-8711-c1891c4e0f47

+++ src/com/intel/bluetooth/BluetoothStreamConnectionNotifier.java
@@ -32,7 +32,7 @@ public class BluetoothStreamConnectionNotifier implements StreamConnectionNotifi
 	
 	private int socket;
 
-	private int handle;
+	private long handle;
 
 	ServiceRecord serviceRecord;
